refactor: Consolidate URLSession protocols into URLSessionProviding.swift
## Summary
- Move `URLSessionBytesProviding` and `URLSessionDataProviding` from individual service files into a single shared `URLSessionProviding.swift`, eliminating parallel-definition risk and making both protocols discoverable in one place
- Fixes #441

## Changes
- Add `RSSApp/Services/URLSessionProviding.swift` with both protocol declarations, the `URLSessionBytesProviding` default-delegate extension, and both `URLSession` conformances
- Remove `URLSessionBytesProviding` protocol block from `ClaudeAPIService.swift`
- Remove `URLSessionDataProviding` protocol block from `GeminiModelService.swift`

## Test plan
- [x] Built successfully for iOS 26 (no errors or warnings)
- [x] No behavior change — pure refactor moving declarations to a shared file
